--- changeFlag/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "change-flag",
-  "version": "1.0.0",
-  "description": ""
-}
