+++ disableSubject/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "change-flag-subject",
+  "version": "1.0.0",
+  "description": ""
+}
